feat: Enable training action - Now we can train combat skills, or if the training station allows, train any other skill, following the definition of data folder structure

+++ idle-mmorpg-server/Core/Factory/Skill/SkillFactory.cpp
@@ -49,7 +49,7 @@ std::unique_ptr<Model::Skill> SkillFactory::createSkill( const std::string& skil
 
             milestoneBonus.setType( bonusJson[ "type" ].asString() );
             milestoneBonus.setId( bonusJson[ "id" ].asString() );
-            milestoneBonus.setValue( bonusJson[ "value" ].asInt() );
+            milestoneBonus.setValue( bonusJson[ "value" ].asDouble() );
             milestone.addBonus( milestoneBonus );
         }
